requirements.txt
asgiref==3.4.1
black==21.11b1
click==8.0.3
colorama==0.4.4
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
drf-nested-routers==0.93.4
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.1
